ci: Use reduced smoke testing on Windows & macOS targets
