+++ sdk/go/common/workspace/plugins_install_nodejs_test.go
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 //go:build nodejs || all
-// +build nodejs all
 
 package workspace
