+++ sdk/go/common/workspace/plugins_install_python_test.go
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 //go:build python || all
-// +build python all
 
 package workspace
