+++ sdk/go/common/workspace/plugins_install_test.go
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 //go:build nodejs || python || all
-// +build nodejs python all
 
 package workspace
