+++ tests/integration/aliases/aliases_dotnet_test.go
@@ -1,6 +1,5 @@
 // Copyright 2016-2020, Pulumi Corporation.  All rights reserved.
-//go:build dotnet || all
-// +build dotnet all
+//go:build (dotnet || all) && !smoke
 
 package ints
