+++ tests/integration/aliases/aliases_go_test.go
@@ -1,6 +1,5 @@
 // Copyright 2016-2020, Pulumi Corporation.  All rights reserved.
-//go:build go || all
-// +build go all
+//go:build (go || all) && !smoke
 
 package ints
