+++ tests/integration/aliases/aliases_nodejs_test.go
@@ -1,6 +1,5 @@
 // Copyright 2016-2020, Pulumi Corporation.  All rights reserved.
-//go:build nodejs || all
-// +build nodejs all
+//go:build (nodejs || all) && !smoke
 
 package ints
