+++ tests/integration/aliases/aliases_py_test.go
@@ -1,6 +1,5 @@
 // Copyright 2016-2020, Pulumi Corporation.  All rights reserved.
-//go:build python || all
-// +build python all
+//go:build (python || all) && !smoke
 
 package ints
