+++ tests/integration/read/read_replace/read_dbr_test.go
@@ -1,6 +1,5 @@
 // Copyright 2016-2018, Pulumi Corporation.  All rights reserved.
-//go:build nodejs || all
-// +build nodejs all
+//go:build (nodejs || all) && !smoke
 
 package ints
